refactor: organize `utils/fetcher` namespace
