+++ src/tts/LingvaTTS/index.ts
@@ -1,5 +1,5 @@
-import { Fetcher } from '../../utils/Fetcher';
-import { basicFetcher } from '../../utils/Fetcher/basicFetcher';
+import { basicFetcher } from '../../utils/fetcher/basicFetcher';
+import { Fetcher } from '../../utils/fetcher/types';
 
 import { TTSAudioBuffer, TTSProviderProps } from '../types';
